FIX Now default image size will display in IE8 and lower
Currently this module doesn't work with IE8 and lower as they do not support media queries so each size with "data-media=" will not load. This adds an IE conditional line to the template which will display the default size. This technique was demonstrated on the picturefill.js docs https://github.com/scottjehl/picturefill#supporting-ie-desktop