Work with higher level resize methods instead of getFormattedImage
#17 raises an issue where a generate*Image method in a module has arguments
that differ to the template-facing API, causing incorrect behaviour when
used in a responsive image set. The arguments for the image resize methods
exposed in templates ($Image.CroppedImage, $Image.PaddedImage, etc) are more
appropriate for configuration via YAML as their parameters must be scalar
to be written in a template.

This change has the bonus of keeping the configuration API for this module
consistent with use of image resize methods in templates, rather than the
potentially different internal API of an image resize method.